feat : 태그 저장 로직 완성 / 노출 미완성

+++ src/main/java/com/encore/bbs/board/dto/BbsDTO.java
@@ -8,7 +8,7 @@ import lombok.*;
 @NoArgsConstructor
 public class BbsDTO {
 
-    private int bbsId;
+    private Long bbsId;
     private int memberCode;
     private String title;
     private String contents;
